// examples/orpc-basic/src/generated/@ir-kit/openapi-ts-typia.gen.ts
export const tAddPetData: StandardSchemaV1<Omit<AddPetData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: AddPetData['path'];
    body: AddPetData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: AddPetData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: AddPetData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}, Omit<AddPetData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: AddPetData['path'];
    body: AddPetData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: AddPetData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: AddPetData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}> & ((input: unknown) => IValidation<Omit<AddPetData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: AddPetData['path'];
    body: AddPetData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: AddPetData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: AddPetData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>) = createValidate<Omit<AddPetData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: AddPetData['path'];
    body: AddPetData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: AddPetData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: AddPetData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>();

export const tAddPetDataJsonSchema = typiaSchemas.schemas[0];

export const tAddPetResponse: StandardSchemaV1<AddPetResponse, AddPetResponse> & ((input: unknown) => IValidation<AddPetResponse>) = createValidate<AddPetResponse>();

export const tAddPetResponseJsonSchema = typiaSchemas.schemas[1];

export const tAddPetResponseError400: StandardSchemaV1<AddPetErrors[400], AddPetErrors[400]> & ((input: unknown) => IValidation<AddPetErrors[400]>) = createValidate<AddPetErrors[400]>();
// ...
export const tFindPetsByStatusData: StandardSchemaV1<Omit<FindPetsByStatusData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: FindPetsByStatusData['path'];
    body: FindPetsByStatusData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: FindPetsByStatusData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: FindPetsByStatusData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}, Omit<FindPetsByStatusData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: FindPetsByStatusData['path'];
    body: FindPetsByStatusData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: FindPetsByStatusData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: FindPetsByStatusData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}> & ((input: unknown) => IValidation<Omit<FindPetsByStatusData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: FindPetsByStatusData['path'];
    body: FindPetsByStatusData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: FindPetsByStatusData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: FindPetsByStatusData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>) = createValidate<Omit<FindPetsByStatusData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: FindPetsByStatusData['path'];
    body: FindPetsByStatusData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: FindPetsByStatusData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: FindPetsByStatusData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>();

export const tFindPetsByStatusDataJsonSchema = typiaSchemas.schemas[9];

export const tFindPetsByStatusResponse: StandardSchemaV1<FindPetsByStatusResponse, FindPetsByStatusResponse> & ((input: unknown) => IValidation<FindPetsByStatusResponse>) = createValidate<FindPetsByStatusResponse>();
// ...
export const tGetPetByIdData: StandardSchemaV1<Omit<GetPetByIdData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: GetPetByIdData['path'];
    body: GetPetByIdData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: GetPetByIdData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: GetPetByIdData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}, Omit<GetPetByIdData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: GetPetByIdData['path'];
    body: GetPetByIdData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: GetPetByIdData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: GetPetByIdData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}> & ((input: unknown) => IValidation<Omit<GetPetByIdData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: GetPetByIdData['path'];
    body: GetPetByIdData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: GetPetByIdData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: GetPetByIdData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>) = createValidate<Omit<GetPetByIdData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: GetPetByIdData['path'];
    body: GetPetByIdData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: GetPetByIdData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: GetPetByIdData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>();
// ...
export const tCreateShapeData: StandardSchemaV1<Omit<CreateShapeData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: CreateShapeData['path'];
    body: CreateShapeData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: CreateShapeData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: CreateShapeData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}, Omit<CreateShapeData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: CreateShapeData['path'];
    body: CreateShapeData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: CreateShapeData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: CreateShapeData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}> & ((input: unknown) => IValidation<Omit<CreateShapeData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: CreateShapeData['path'];
    body: CreateShapeData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: CreateShapeData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: CreateShapeData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>) = createValidate<Omit<CreateShapeData, 'url' | 'path' | 'body' | 'query' | 'headers'> & {
    params: CreateShapeData['path'];
    body: CreateShapeData extends {
        "body": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    query: CreateShapeData extends {
        "query": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
    headers: CreateShapeData extends {
        "headers": infer V;
    } ? [
        V
    ] extends [
        undefined
    ] ? unknown : V : unknown;
}>();